signals: Generate signals into pyi files as attributes
This patch adds signals as normal attributes to pyi files.
It may be replaced by a better version, later:

It is also possible to create them as functions which document
the signature of the creating signal call. This only makes sense
when the signature support in C++ is extended for signals.

Change-Id: Id25cc93e905ea697c23b27ec078e0ff997814624
Fixes: PYSIDE-1603
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>