Fix debug build problems on Win re: limited-api
There is a build error on Windows when limited-api=yes and debug where
the linker does not know the right paths to look for libraries.
Supporting this configuration is not desirable anyway because there are
no released Python builds on Windows with this configuration. Therefore:

1. Display a warning to the user that the combination debug +
   limited=api=yes does not really "exist" if they choose it explicitly
   on Windows.
2. Have the default for debug builds be limited-api=no on Windows.

Change-Id: Ifed99f80abb02c9d512a1321526e949d56ff08a6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 72107fc39f15e3827297530dfad3c63f8e4ffcaa)