signature: Fix handling of rlcompleter as intermediate solution
The rlcompleter module calls inspect.signature which breaks
the multi-signatures of the signature C module.

The __signature__ attribute should no longer be used anyway
and people should use get_signature(), instead. But before we
evict this attribute, this patch fixes the problem, temporarily.

[ChangeLog][PySide6] Python 3.10 and up have an incompatibility
to the __signature__ attribute. This temporary patch fixes this
until final removal of that attribute.

Change-Id: Ic08d551cd388e6b3b29eb4694427a7202a4272b4
Fixes: PYSIDE-2101
Pick-to: 6.4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>