__feature__: heavily rework the context switching
The example of the issue shows the qasync.py module which
fails miserably when using snake_case.

The reason:
-----------
Reason is the way how feature switching is implemented.
Modules like qasync get a default switching of "ignore".
This ignores that the qasync module itself imports QtCore,
and feature switching is of course relevant, suggesting
a default setting of "normal" (explicitly no features).

The real problem:
-----------------

Testing the simple approach showed a serious problem with
feature switching: The functions get switched when a certain
function (getattr etc.) is called.
But the switching is sometimes not done due to a caching problem.

This fix removes caching that was wrong. Optimization will
be done in a different step with a different approach.
This Change was not qasync specific, but happens with PySide imports.

Actions done:
- adjust the inline structure
- implement a feature_imported callback
- identify Python functions that use PySide during import

[ChangeLog][PySide6] __feature__ switching now works even with
                     recursive imports like in the qasync module.

Fixes: PYSIDE-2029
Change-Id: I3340f54f293083a09fb509383688f73bbd9b60ae
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 7377d2b8130ce7290775cd8a343e75c0561fc854)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>