shiboken6: Generate property documentation
Previously, the documentation generator used to insert
the property documentation at the getter/setter function documentation.

The properties for use with true_property are hard to discover in this
scheme.

To fix this, add a separate TOC section and description list
for the properties using the sphinx :py:property: directive
and link to it from the functions and signals instead.
This mimicks the C++ documentation.

Task-number: PYSIDE-1106
Task-number: PYSIDE-1019
Pick-to: 6.4
Change-Id: I976fee91a02ca4c8a7c62c7d957ecaea59ac4ebc
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>