Get rid of PyErr_Print in callbacks
PyErr_Print is used since a pretty long time (2010)
to handle errors which cannot be treated normally
in callbacks.

This is annoying since it happens in GUI code, most
of the time, and it is impossible to catch an exception.

This patch delays the error handling until we are in
a normal Python context.

Example of the PYSIDE-1889 error:

"""
RuntimeError: Delayed RecursionError exception:

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/Users/tismer/src/QtC/pyside-setup/pyside1889.py", line 23, in <module>
    app.exec()
RecursionError: maximum recursion depth exceeded while calling a Python object
"""

In a traceback, the recursion error is correctly reported as top.

[ChangeLog][PySide6] An exception that cannot be raised is no longer
simply printed, but re-raised as an exception as soon as possible.

Change-Id: I9f37de745d9025b38bd3a9fa523c34ea88c583c3
Task-number: PYSIDE-1889
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>