build_system: Remove code trying to find the MSVC installation
The code was not triggered since the build system tries to compile
a small test project with Ninja before even reaching it.

It was also using code for MSVC 9 (2008) which does not work anymore
and causes deprecation warnings:

setuptools\_distutils\msvc9compiler.py:34: DeprecationWarning:
msvc9compiler is deprecated and slated to be removed in the future.
Please discontinue use or file an issue with pypa/distutils describing
your use case.

Change-Id: I95a6b7700c13873884deb6186e78d6c520c592de
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Adrian Herrmann <adrian.herrmann@qt.io>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>