shiboken6: Include base class wrapper headers of the current module when avoiding the protected hack
When avoiding the protected hack, the generated wrapper headers
need to be included to generate casts to the wrapper classes.

This was previously done by generating the entire class declaration
into the header, omitting its extra includes, which causes clashes in
CMake UNITY_BUILD (jumbo) builds and is a violation of the ODR rule.

Improve this by generating include statements for inherited
wrapper headers of the current module.  For classes from other modules,
generate the declarations as before, as wrapper headers
are not installed.

Task-number: PYSIDE-2155
Task-number: PYSIDE-500
Change-Id: Ib7caa3f49901090d7d94274018e04ccde2ebfdf9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>