build_scripts: Change source of bundled MSVC DLLs
Currently, PySide bundles MSVC DLLs (+ others) sourced from a very
outdated self-hosted archive. Best practice would be relying on the
redistributable package and installing it if necessary, but we cannot
assume that the user has the required admin rights. As a compromise, the
bundled DLLs are now sourced from the CI machine, which is assumed to be
kept somewhat up-to-date, at least on a better cadence than the current
3+ year old archive.

These non-MSVC DLLs are no longer bundled as they are pre-installed on
Windows: d3dcompiler (part of DirectX) and ucrtbase (Microsoft C Runtime
Library). Only opengl32sw continues to be sourced from the self-hosted
archive.

Pick-to: 6.2 5.15
Task-number: PYSIDE-2162
Change-Id: I6495f6953cd1c3e78a9777d24ee54e429a63d3d9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>