property/QML: Improve handling of an unknown class name
pysideproperty has Q_ASSERT calls, which cause errors when
a typename is undefined. This behavior is inconsistent and
should either raise an exception or be circumvented.

For compatibility with another implementation, unknown
types are now ignored, and debug and release are consistent.

To obtain an error, run the script with -Werrors . This makes
use of the new exception delay feature.

Task-number: PYSIDE-2160
Change-Id: Ia320edab4b51dd6f6b94a429b8f1f086ce9897f7
Pick-to: 6.4 6.2
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>