feature: Make _mod_uses_pyside safe against binary modules
Investigation whether a module uses PySide or not requires
access to the source code of the module.
This is at least true for our purpose.

But inspect.getsource(module) seems to try to get source
code even if this is a binary module, and then fails
with a unicode error.

Change-Id: I87fe9a0aae21318581513f6ce2b0c903a6b82e6b
Fixes: PYSIDE-2189
Pick-to: 6.4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>