shiboken: defer the feature initialization to be safe in other contexts
The initialization of the feature import re-defines certain import hooks
during the initialization phase in loader.py.

This does currently only work because all Python modules are pre-
loaded and therefore already in memory. When de-virtualizing the
Python modules, a real import happens, and we get a timing problem.

Solved by delaying the initialization of feature imports until the
functions of signature_globals.cpp are initialized.

Task-number: PYSIDE-1019
Change-Id: If8fee08b20e4b25eaf1719f03189e3de04ace8c5
Pick-to: 6.4
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>