shiboken: optionally de-virtualize the Python files
The Python files in Shiboken are completely virtualized. That
means: There is no interaction with the external world that
could create interferences.

For debugging purposes or to modify the embedded files, it
is anyway useful to have an option to materialize these files
in the file system, again.

We add an environment variable `SBK_EMBED` that explicitly
unpacks the Python files (SBK_EMBED=0) or explicitly removes
the unpacked files (SBK_EMBED=1).

Without this variable, the state of embedding remains the same.

Change-Id: I1081b0f910482fb4b80a02f72c6bcce060db38e6
Fixes: PYSIDE-1994
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit ee35c071ed8493cc5e3a80f19d9e2c2348160d56)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>