create_wheels: Add --build-dir argument
If PySide was built outside a venv, the build directory has a different
naming scheme and --env as the name of the option is a bit unintuitive.
Add a --build-dir option to directly provide the relative path of the
build directory. Additionally, if no valid build directory or env name
were given, make it clearer to users that create_wheels is using the
first valid directory it could find, which might not be the wanted one.
Both --env and --build-dir options are now better explained upon running
--help.

Pick-to: 6.4
Change-Id: Idac9d48c85b2d0b5611d3f28f0f04da5ec7e3d3a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>