nogil: Adjust the forgiveness enum opcode patch
The forgiveness mode has opcode introspection for
determining the necessary replacement strategy.

The current NoGIL code has very different opcodes.
Not even the long introduced word addressing is used,
and it is now a register machine.

This implementation will vanish, soon. But the effort
was affordable to support. The real problem of
supporting NoGIL stays the same.

Task-number: PYSIDE-2221
Change-Id: Ida2dd38b042f35ab5b1bc0ea232d5241d7176c0a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>