Python 3.12: Fix the type generation intermediately
The type generation that we use all over the place does not work
any longer because Python 3.12.0a5 no longer allows meta classes with
a tp_new function.

The developers are aware of the omission and agree that a fix must
be found. After prior experience, this can take very long time.
To immediately continue, we implement a temporary patch that works.

Task-number: PYSIDE-2230
Change-Id: I8bf237411ca0e32e31ef4d3741a2def05f3e39f6
Pick-to: 6.4 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>