doc: migrate to sphinx-design and update sphinx
sphinx-panels is deprecated, and sphinx-design is the better
replacement.

This patch migrates everything we had in pyside and shiboken,
which was useful to remove many CSS related hacks for things
we were forcing before.

The example gallery was also adapted to generate sphinx-design
valid code. Additional adaptation to some examples and files were
included as well, which improved the overall example experience.

We used this to change to update sphinx as well, which before
was not possible to update due to sphinx-panels.

Fixes: PYSIDE-2244
Change-Id: Ia3f57f536591ea299bf87302e930203b64fafd7e
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>