shiboken: Fix getOverride to support private methods
Private methods generate no wrappers, and so the logic in
BindingManager::getOverride does not work.
Return the method as-is in this case.

An acknowledgment goes to Friedemann for the right idea.
And a -2 for Gerrit's pickiness :)

Fixes: PYSIDE-2255
Change-Id: If1f0bcc037c7c190eaedab1a245b35062c426783
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 10b3c162647d0d25e34bf1e3ae5b2ce57c89f324)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>