PyPySide: Fix a crash with PyPy 7.3.10 and 7.3.11
The PyPy installation had no problems with the built-in _functools module, which
we still use in favor of the official functools module, which is slower and can be
manipulated. The _functools module is no longer found in PyPy, and speed issues
do not apply.

[ChangeLog][shiboken6] A PyPy problem with version 7.3.10 and up was fixed.

Change-Id: Ie064d803d3fbf0f4ca226329b4d317db2a44c045
Fixes: PYSIDE-2264
Task-number: PYSIDE-535
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 3b91e4a7886c8ca286bd50cbf638de15ff386919)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>