doc: use new logo and colors
This updates the branding of the docs to the new guidelines
in Qt.

The neon green was updated to #2CDE85,
and for the light version of the docs, because of the contrast
Midnight was used #27138B (Pine, and Moss didn't work well)

The Qt for Python logo was updated as well.

Change-Id: I2cf556d5c3b3c9c84923204f3c2e829ab1f68f54
Pick-to: 6.5
Reviewed-by: Adrian Herrmann <adrian.herrmann@qt.io>
Reviewed-by: Jaishree Vyas <jaishree.vyas@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>