shiboken6: Refactor code snippet extraction
The support for doc snippets with '#' as used in CMakeLists and .pro
files was only partially working; the search string did not match.

Refactor the code to use only one regex that can match both and
extracts the id.

Add an error message for empty snippets as are produced by
snippets_translate by occasionally repeating snippet ID lines.

Add a test.

Task-number: PYSIDE-1106
Pick-to: 6.5
Change-Id: I77d465ff52f78d6ad6286e83b5c05015e6ea0e46
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>