Fix leak connecting signals to anonymous functions
When connecting a signal to the same anonymous function repeatedly,
a new GlobalReceiverV2 object would be created after each call. Each
GlobalReceiverV2 would have a unique callback object despite all these
callback objects sharing the same code. This would lead to a large
number of GlobalReceiverV2 and callback objects, each never reaching a
refcount of 0 and thus never being released. The remedy is that we only
need one GlobalReceiverV2 object, whose corresponding GlobalReceiverKey
references not the outer callback object, but the code object associated
with it.

Pick-to: 6.5
Fixes: PYSIDE-2299
Change-Id: I474284dc5ce08dc6601636f2e7ac5e5a10ed8560
Reviewed-by: Christian Tismer <tismer@stackless.com>