shiboken: Implement raising of unraisable exceptions correctly
TODO in another check-in: Control exec/run by an XML attribute

[ChangeLog][PySide6] Unraisable exceptions are now handled by a
                     handler on the stack or printed if impossible.

Unraisable exceptions are stored in an error store if there is
an error handler on the call stack that can handle it,
otherwise it is printed immediately.

We record the existence of an error handler in thread local
storage, which solves thread problems automatically.

Since exec and run functions completely block all handlers,
we need to mark them as a special case.

The overhead is minimal and uses constant memory per thread.

Task-number: PYSIDE-2310
Task-number: PYSIDE-2321
Change-Id: Ic25a2ff8552baf6e132ad86a4ad0925375e7ea88
Fixes: PYSIDE-2335
Pick-to: 6.5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>