Shiboken: Undo the unraisable exception delay for now
The delayed exception looked good in the first place.

PYSIDE-2310 and PSIDE-2321 showed some special cases where
the delayed error reporting fails to notify about an error,
to be in time (2321) or all-together (2310).

This is no complete reversal of the thing. We need to
improve the theory to cover all cases and do a better
implementation.

This temporary reversal removes the pressure from us to
get the feature perfect before the 6.5.1 release.

Change-Id: Idf25ca31571a45c59082d27e4d736a4932ab9f8c
Fixes: PYSIDE-2310
Fixes: PYSIDE-2321
Pick-to: 6.5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>