doc: clarify issue with Qt installations
This arose from many similar comments from people
new to Python, and wondering about the role of having
Qt installed in the system.

Additionally, some of the Qt-classes references in the file
were not linked to the actual documentation page, so the usage
of :ref: rather than :code: is encouraged.

Pick-to: 6.5
Change-Id: I3d5560a49b7a89507aa8774771f4a6acacb67d6b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>