Revert anonymous slot memory leak fix
The fix to PYSIDE-2299 unfortunately introduces regressions that are
more serious than the memory leak caused by an edge case that was
originally supposed to be fixed. A different approach will be needed to
fix the memory leak, in the meantime, revert the change. The
corresponding unit test remains blacklisted, as it will apply to a
future fix.

Pick-to: 6.5
Fixes: PYSIDE-2346
Task-number: PYSIDE-2299
Change-Id: I1dcddb03f731a333f6acf7c9df79b9dba42bec21
Reviewed-by: Christian Tismer <tismer@stackless.com>