build: remove traces of distutils
Removed mentions, comments, and fallback usages
in some sections of the build scripts and documentation.

On the CI, we upgrade setuptools in order to avoid having
missing functionality, so it shouldn't be an issue.

For Shiboken and CMake, there is still a helper that relies
on distutils in case people is using python < 3.8.2

Pick-to: 6.5
Change-Id: Ic33abe9fa8bd8564e9d44c8714b7fedd4e50434a
Reviewed-by: Simo Fält <simo.falt@qt.io>