build: remove traces of distutils
Removed mentions, comments, and fallback usages
in some sections of the build scripts and documentation.

On the CI, we upgrade setuptools in order to avoid having
missing functionality, so it shouldn't be an issue.

For Shiboken and CMake, there is still a helper that relies
on distutils in case people is using python < 3.8.2

Change-Id: Ic33abe9fa8bd8564e9d44c8714b7fedd4e50434a
Reviewed-by: Simo Fält <simo.falt@qt.io>
(cherry picked from commit 709c3f0b8a7a597e531506d6024ead26b3a8f08a)
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>