PySide6/SignalManager: Fix memory leaks connecting free methods/lambdas
Change 1270a9e82e5bc3bd53a1131698ece60403da1192 changed the deletion
of global receivers from listening to QObject::destroyed() (which
caused thread issues) to using QPointer<> and purging the lists in
notify. What is missing was the deletion of global receivers that are
not tied by weak reference to a Python instance. Add a check in
notify() to clean out the empty global receivers.

Pick-to: 6.5
Fixes: PYSIDE-2371
Fixes: PYSIDE-2299
Task-number: PYSIDE-2141
Change-Id: I39dca2a21088930c9a7f8e5eb7e948b3fff49b4b
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>