Fix stub and signature for QClipboard setPixmap and setImage
Incorrectly, the pyi file described that it was possible to pass a:
    Union[PySide6.QtGui.QPixmap, PySide6.QtGui.QImage, str]
for setPixmap, and:
    Union[PySide6.QtGui.QImage, str]
for setImage.

This uses the typesystem entry to restrict the setPixmap
to accept a QPixmap and only a QImage for setImage.

Additionally, this enable a PyPathLike signature, so users
can pass 'str', 'bytes', 'pathlib.Path' as well, so the accepted
arguments are:

  setImage(...)
    - PySide6.QtGui.QImage
    - Union[str, bytes, os.PathLike]

  setPixmap(...):
    - PySide6.QtGui.QPixmap
    - Union[str, bytes, os.PathLike]

Pick-to: 6.5
Fixes: PYSIDE-2395
Change-Id: I521980dbc781a72e54f49fc84af5349733385133
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>