Handle more attributes in add-function/declare-function/function elements
Factor out functions to handle the attributes from modify-function and
use them for <add-function>/<declare-function> and <function>.

This makes it possible to specify the "allow-thread",
"exception-handling" and "snake-case" for <declare-function> and
<function> as well.

The snake-case handling is removed from FunctionTypeEntry and handled
via modifications.

As a drive-by, fix docs on "exception-handling".

[ChangeLog][shiboken6] The attributes "allow-thread",
"exception-handling" and "snake-case" can now be specified on
<declare-function> and <function>.

Change-Id: I081ab4dfd922de563ac7b8c75d0c36609c43e0cd
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>