Fix QLocale.toLong return value
The currently implementation was using an inherited toLong
function, not returning the Tuple[int, bool] that the docs
described. This adds a modification to follow the same
idea of all the other to*() methods.

Additionally, we remove all the to*() methods that
accept a QStringView, due to behaving the same with the QString
variations.

[ChangeLog][pyside6] QLocale.toLong now properly returns a (int, bool) rather than only int

Fixes: PYSIDE-2226
Change-Id: I5634e4010982e115c8208fdb9b56cfc57960358b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>