libpyside: Refactor passing slot data between @Slot and MetaObjectBuilder
The slot data required for MetaObjectBuilder were stored in a PyList
set as an attribute on a method. This required concatenating return
type and signature, converting the resulting string to CPython in the
@Slot code and converting it back to QByteArray in MetaObjectBuilder.

To simplify this, introduce a small C++ struct storing the QByteArrays
and store a list of them as a PyCapsule (encapsulating a void *ptr)
which is used as the attribute value.

Task-number: PYSIDE-748
Change-Id: I7f4075f5e828fe543d01e5dfbdc7087905cd004f
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>