PyPySide: Fix a small, very old glitch
Many years ago (2012), the format string of
SbkObjectType_tp_new contained the string "SO!O!:sbktype".
The meaning of this "S" is redundant for historical reasons
and means PyObject. It is equal to "O".

At some unknown point in time, this "S" was erroneously
turned into "s", which means a plain string.

Unfortunately, this neat finding does not solve the multiple
inheritance problem of PyPy, but this will surely be found
at some point, too.

Change-Id: I51045216f9c0aa2c9efdda06a2ad5f9ac64e9696
Task-number: PYSIDE-535
Pick-to: 6.5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>