shiboken6: Make generator options statically accessible
Create struct withs options (to be used by an improved command line
parser later on) and make them static members of the generators.

This is a first step to remove the currently duplicated handling of
options since ShibokenGenerator is instantiated for HeaderGenerator
and CppGenerator.

Also, more generator functions can then be made statically accessible.

Pick-to: 6.6 6.5
Change-Id: I1e355370ef32d35ddd9f1e7d847a1965eb44e077
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Adrian Herrmann <adrian.herrmann@qt.io>