QML registration code: Extract a setter from the ClassInfo decorator
Change it to operate on a list of key/value pairs instead of a QMap
(no use in enforcing sorting) and extract a setter for setting
QMetaClassInfo data on a PyTypeObject of a QObject.
To be used for QML decorators.
As a drive-by, fix some static analysis warnings about implicit bool
casts.

Task-number: PYSIDE-2484
Change-Id: Ia759bb42740ed279f36c0850306ebd9bee526ecf
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>