CMake: Allow specifying a shiboken debug-level from the build system
The SHIBOKEN_DEBUG_LEVEL CMake or env var variables can now be set so
that the build system passes that value to --debug-level of each
shiboken invocation.

The possible values are the ones that shiboken's command line parser
expects: sparse, medium, full.

This is useful for getting shiboken debug output in the CI.

Pick-to: 6.6
Change-Id: I62d99e8f67884638449a69b10803aca3b1230e7b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>