Support running PySide on Python 3.12, cleanup
In the course of fixing 3.12 issues, a 3.11 fix for PyEnum
was missing that created a 3.11 error.

A bug with deployment was fixed but not reflected in the tests.
This was only visible for pyenv users.

Also fixes a minor cosmetic bug introduced by
"Python 3.12: Fix the structure of class property"

Task-number: PYSIDE-2230
Change-Id: I3a6bd6426e23168dfccdda17b408f193aacd28a0
Pick-to: 6.5 6.2
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 29967cb08e6dfd19b0e49013d3b936f9fd18c6a2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>