Support running PySide on Python 3.12, extension bug
When SbkTypeFactory._PyType_FromSpecWithBasesHack was
written, the intent was to support everything for
Python 3.12, although the new function PyType_FromMetaclass
cannot be used with the Limited API.

The supplied patch did not fix a problem that arises when
a type is created without a base class that has the preferred
metaclass. In PySide, this happens exactly in the case of
SbkObject_type creation.

The subtle bug was that metaclass SbkObjectType_Type was
reported, but it really was still the old &PyType_Type.
The patch fixes that by temporarily changing the type of object.

Python 3.12 is the first version that requires that.

Change-Id: I4d9d05ed93aa0e1fb297a0b931821ab9cae28ee9
Pick-to: 6.6 6.5 6.2
Task-number: PYSIDE-2230
Reviewed-by: Adrian Herrmann <adrian.herrmann@qt.io>