Create PySide6_Examples wheel
This adds the functionality of creating a wheel that
only will contain the examples directory from the repo
and place it under an existing 'PySide6' diretory
in the site-packages.

The dependencies of the wheel are the same as the empty
'PySide6' wheel, meaning PySide6_Essentials and PySide6_Addons
will be installed if not present.

Please consider that the 'hacks' we are using to create
the wheels in our current project structure requires
for us to have a combination of the files:
- setup.py
- setup.cfg
- pyproject.toml
- MANIFEST.in

Without a proper directory structure per project,
it's not possible to drop a couple of them.

Pick-to: 6.6
Change-Id: I164abb30bb744a98ec1d31b304a0a3668a543ae2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Adrian Herrmann <adrian.herrmann@qt.io>