Introduce a callback functor for QML singleton creation
Add a functor class with shared data that keeps a reference to an object.
Preparing for handling a static create() function as a singleton
creation callback, this allows for clean separation of the
object to keep the reference and the object to be called
and provides a way of cleaning up. The cleanup cannot be currently
activated since QML registration data are kept in global variables.

Task-number: PYSIDE-2432
Pick-to: 6.6
Change-Id: Id57811316e8803638ac3148fdad18a854be99cca
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>