Shiboken: Fix subtle bug caused by PyPy support and "bug_825"
When implementing PyPy support, all Shiboken types lost their
hidden extra fields, because they were replaced by shadow dicts.

This creates a problem with multiple inheritance.
In Python < 3.12, the decision which base type will create the
instance is driven simply by the size of the type. In

    class MetaC(MetaA, MetaB):
        pass

when MetaA is derived from type and MetaB is derived from
SbkObjectType, MetaA is not the type that should win.
Fixed by simply adding 1 to the default size of PyType_Type.

Caused by bug_825 corrected version.
The old test is retained as bug_825_old.py .

NOTE: This touches a Python 3.8 bug that was fixed in 3.9.12
and 3.10.4 - unfortunately CI Python is older in all cases.

Task-number: PYSIDE-2230
Change-Id: I6e82cafb83c8351bbbeafcc80e11d5e45568f73d
Pick-to: 6.2 6.5 6.6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>