testing: Improve the Python version info to get better control
The QML bug that was identified by "bug_825" revealed
a Python error which is persistent in Python 3.8 and
was fixed in Python 3.9.12 and Python 3.10.4 .

It was not possible to write a work-around without
re-implementing large areas of the Python type system.

We interrogate the TestRunner to obtain the full Python
version. Otherwise we would have to exclude Python 3.8
to 3.10 completely.

Task-number: PYSIDE-2230
Change-Id: Ica53c2e7b44cbbf5ec8ca1430ab65e6743beeff8
Pick-to: 6.6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>