QTimer: Implement singleShot signatures w/ context
On C++, it is possible to call singleShot timers with a context object
as an argument. This allows posting events to the event loop of the
given context object's thread, instead of the thread of the current
thread. Implement corresponding signatures to add this capability to
Qt for Python.

Change-Id: I0c4e3ef4b859cdfaac07415ff64c440821e7f442
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 671c182dce91c387037894017be314564e054c00)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>