PySide6: Fix obtaining properties from QDBusInterface
Disable the PySide6 meta object functions of QDBusInterface since they
clash with the dynamic meta object builder it uses.

Note that it is then no longer possible to derive a Python
class with signals/slots from QDBusInterface.

[ChangeLog][PySide6] It is no longer possible to derive a Python class
with signals/slots from QDBusInterface. This is necessary to able to
retrieve properties via DBus.

Fixes: PYSIDE-2535
Change-Id: I6622c7d0bbc8e2684d1390e2888dc68053464189
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>